feat: Refactor project file tree
