+++ src/Requests/PurchaseEvent.php
@@ -6,12 +6,12 @@
  * @modifiedBy <name>
  */
 
-namespace ZaiKorea\ZaiClient\Requests;
+namespace ZaiClient\Requests;
 
-use ZaiKorea\ZaiClient\Requests\BaseEvent;
-use ZaiKorea\ZaiClient\Requests\EventInBatch;
-use ZaiKorea\ZaiClient\Configs\Config;
-use ZaiKorea\ZaiClient\Exceptions\BatchSizeLimitExceededException;
+use ZaiClient\Requests\BaseEvent;
+use ZaiClient\Requests\EventInBatch;
+use ZaiClient\Configs\Config;
+use ZaiClient\Exceptions\BatchSizeLimitExceededException;
 
 /**
  * @final
@@ -99,13 +99,15 @@ class PurchaseEvent extends BaseEvent
             }
 
             for ($i = 0; $i < $order_spec['count']; $i++) {
-                array_push($events, new EventInBatch(
-                    $user_id,
-                    $order_spec['item_id'],
-                    $tmp_timestamp,
-                    self::EVENT_TYPE,
-                    $order_spec['price']
-                )
+                array_push(
+                    $events,
+                    new EventInBatch(
+                        $user_id,
+                        $order_spec['item_id'],
+                        $tmp_timestamp,
+                        self::EVENT_TYPE,
+                        $order_spec['price']
+                    )
                 );
                 $tmp_timestamp += Config::EPSILON;
             }
